packages/cowork-core/pyproject.toml
[project]
name = "cowork-core"
version = "0.1.0"
description = "Cowork core — ADK agents, tools, policy, session, memory, workspace"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.2",
    "litellm>=1.40",
    "pydantic>=2.8",
    "httpx>=0.27",
    "ddgs>=0.9",
    "pyyaml>=6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cowork_core"]
